objects/pickup: merge setup and control in TRX
