+++ src/libtrx/game/objects/general/pickup.c
@@ -1,7 +1,18 @@
 #include "game/objects/general/pickup.h"
 
 #include "config.h"
+#include "game/effects.h"
 #include "game/lara.h"
+#include "game/random.h"
+#include "game/savegame.h"
+
+// clang-format off
+#define M_AID_DIST_MIN          (STEP_L * 5)      // 1280
+#define M_AID_DIST_MAX          (WALL_L * 8)      // 8192
+#define M_AID_WAIT_MIN          (LOGIC_FPS * 2.5) // 75
+#define M_AID_WAIT_MAX          (LOGIC_FPS * 5)   // 150
+#define M_AID_WAIT_BREAK_CHANCE 0x1200
+// clang-format on
 
 static const OBJECT_BOUNDS m_PickUpBounds = {
     .shift = {
@@ -36,6 +47,99 @@ static const OBJECT_BOUNDS m_PickUpBoundsUW = {
     },
 };
 
+static void M_Initialise(int16_t item_num)
+{
+    ITEM *const item = Item_Get(item_num);
+    item->priv = (void *)(intptr_t)(-1);
+    if (item->status != IS_INVISIBLE) {
+        Item_AddActive(item_num);
+    }
+}
+
+static void M_HandleSave(ITEM *const item, const SAVEGAME_STAGE stage)
+{
+    if (stage == SAVEGAME_STAGE_AFTER_LOAD) {
+        if (item->status == IS_DEACTIVATED) {
+            const int16_t item_num = Item_GetIndex(item);
+            Item_RemoveDrawn(item_num);
+        }
+    }
+}
+
+static void M_Activate(ITEM *const item)
+{
+    if (item->status == IS_INVISIBLE) {
+        item->touch_bits = 0;
+        item->status = IS_ACTIVE;
+        const int16_t item_num = Item_GetIndex(item);
+        Item_AddActive(item_num);
+    } else {
+        item->status = IS_INVISIBLE;
+        item->flags |= IF_KILLED;
+    }
+}
+
+static void M_SpawnPickupAid(const ITEM *const item)
+{
+    const OBJECT_ID obj_id =
+        Object_GetCognate(item->object_id, g_ItemToInvObjectMap);
+    const OBJECT *const obj = Object_Get(obj_id);
+    const ANIM_FRAME *const frame = obj->frame_base;
+
+    const GAME_VECTOR pos = {
+        .x = item->pos.x + 20 * (Random_GetDraw() - 0x4000) / 0x4000,
+        .y = item->pos.y - ABS(frame->bounds.max.y - frame->bounds.min.y)
+            - 10 * (1 + (Random_GetDraw() - 0x4000) / 0x4000),
+        .z = item->pos.z + 20 * (Random_GetDraw() - 0x4000) / 0x4000,
+        .room_num = item->room_num,
+    };
+
+    const int16_t effect_num = Effect_Create(pos.room_num);
+    if (effect_num != NO_EFFECT) {
+        EFFECT *const effect = Effect_Get(effect_num);
+        effect->room_num = pos.room_num;
+        effect->pos.x = pos.x;
+        effect->pos.y = pos.y;
+        effect->pos.z = pos.z;
+        effect->counter = 0;
+        effect->object_id = O_PICKUP_AID;
+        effect->frame_num = 0;
+    }
+}
+
+static void M_Control(int16_t item_num)
+{
+    ITEM *const item = Item_Get(item_num);
+    if (item->status == IS_INVISIBLE || item->status == IS_DEACTIVATED) {
+        Item_RemoveActive(item_num);
+        return;
+    }
+#if TR_VERSION == 1
+    const ITEM *const lara = Lara_GetItem();
+    if (!g_Config.gameplay.enable_pickup_aids || item->fall_speed != 0
+        || lara == nullptr || !Object_Get(O_PICKUP_AID)->loaded) {
+        return;
+    }
+
+    const int32_t distance = Item_GetDistance(lara, &item->pos);
+    if (distance < M_AID_DIST_MIN || distance > M_AID_DIST_MAX) {
+        return;
+    }
+
+    int32_t timer = (int32_t)(intptr_t)item->priv;
+    if (timer <= 0
+        || (timer < M_AID_WAIT_MIN
+            && Random_GetDraw() < M_AID_WAIT_BREAK_CHANCE)) {
+        M_SpawnPickupAid(item);
+        timer = M_AID_WAIT_MAX;
+    } else {
+        timer--;
+    }
+
+    item->priv = (void *)(intptr_t)(int32_t)timer;
+#endif
+}
+
 const OBJECT_BOUNDS *Pickup_Bounds(void)
 {
     const LARA_INFO *const lara = Lara_GetLaraInfo();
@@ -63,3 +167,50 @@ bool Pickup_Trigger(const int16_t item_num)
     item->status = IS_DEACTIVATED;
     return true;
 }
+
+static void M_Setup(OBJECT *const obj)
+{
+    obj->activate_func = M_Activate;
+    obj->control_func = M_Control;
+    obj->collision_func = Pickup_Collision;
+    obj->bounds_func = Pickup_Bounds;
+    obj->draw_func = Object_DrawPickupItem;
+    obj->initialise_func = M_Initialise;
+    obj->handle_save_func = M_HandleSave;
+    obj->save_position = true;
+    obj->save_flags = true;
+}
+
+REGISTER_OBJECT(O_EXPLOSIVE_ITEM, M_Setup)
+REGISTER_OBJECT(O_FLARES_ITEM, M_Setup)
+REGISTER_OBJECT(O_GRENADE_AMMO_ITEM, M_Setup)
+REGISTER_OBJECT(O_GRENADE_ITEM, M_Setup)
+REGISTER_OBJECT(O_HARPOON_AMMO_ITEM, M_Setup)
+REGISTER_OBJECT(O_HARPOON_ITEM, M_Setup)
+REGISTER_OBJECT(O_KEY_ITEM_1, M_Setup)
+REGISTER_OBJECT(O_KEY_ITEM_2, M_Setup)
+REGISTER_OBJECT(O_KEY_ITEM_3, M_Setup)
+REGISTER_OBJECT(O_KEY_ITEM_4, M_Setup)
+REGISTER_OBJECT(O_LARGE_MEDIPACK_ITEM, M_Setup)
+REGISTER_OBJECT(O_LEADBAR_ITEM, M_Setup)
+REGISTER_OBJECT(O_M16_AMMO_ITEM, M_Setup)
+REGISTER_OBJECT(O_M16_ITEM, M_Setup)
+REGISTER_OBJECT(O_MAGNUM_AMMO_ITEM, M_Setup)
+REGISTER_OBJECT(O_MAGNUM_ITEM, M_Setup)
+REGISTER_OBJECT(O_PICKUP_ITEM_1, M_Setup)
+REGISTER_OBJECT(O_PICKUP_ITEM_2, M_Setup)
+REGISTER_OBJECT(O_PISTOL_AMMO_ITEM, M_Setup)
+REGISTER_OBJECT(O_PISTOL_ITEM, M_Setup)
+REGISTER_OBJECT(O_PUZZLE_ITEM_1, M_Setup)
+REGISTER_OBJECT(O_PUZZLE_ITEM_2, M_Setup)
+REGISTER_OBJECT(O_PUZZLE_ITEM_3, M_Setup)
+REGISTER_OBJECT(O_PUZZLE_ITEM_4, M_Setup)
+REGISTER_OBJECT(O_SCION_ITEM_2, M_Setup)
+REGISTER_OBJECT(O_SECRET_1, M_Setup)
+REGISTER_OBJECT(O_SECRET_2, M_Setup)
+REGISTER_OBJECT(O_SECRET_3, M_Setup)
+REGISTER_OBJECT(O_SHOTGUN_AMMO_ITEM, M_Setup)
+REGISTER_OBJECT(O_SHOTGUN_ITEM, M_Setup)
+REGISTER_OBJECT(O_SMALL_MEDIPACK_ITEM, M_Setup)
+REGISTER_OBJECT(O_UZI_AMMO_ITEM, M_Setup)
+REGISTER_OBJECT(O_UZI_ITEM, M_Setup)
