+++ src/tr1/game/objects/general/pickup.c
@@ -1,60 +1,20 @@
-#include "game/effects.h"
-#include "game/game.h"
-#include "game/inventory.h"
-#include "game/lara.h"
-#include "game/objects/common.h"
-#include "game/savegame.h"
 #include "game/stats.h"
 
 #include <libtrx/config.h>
-#include <libtrx/game/gun.h>
 #include <libtrx/game/input.h>
-#include <libtrx/game/lara/common.h>
+#include <libtrx/game/inventory.h>
+#include <libtrx/game/lara.h>
 #include <libtrx/game/lua/common.h>
 #include <libtrx/game/lua/events.h>
+#include <libtrx/game/objects.h>
 #include <libtrx/game/objects/general/pickup.h>
-#include <libtrx/game/objects/vars.h>
 #include <libtrx/game/overlay.h>
-#include <libtrx/game/random.h>
 
 #define LF_PICKUP_ERASE 42
 #define LF_PICKUP_UW 18
-#define AID_DIST_MIN (STEP_L * 5)
-#define AID_DIST_MAX (WALL_L * 8)
-#define AID_WAIT_MIN (LOGIC_FPS * 2.5)
-#define AID_WAIT_MAX (LOGIC_FPS * 5)
-#define AID_WAIT_BREAK_CHANCE 0x1200
 
 static XYZ_32 m_PickUpPosition = { 0, 0, -100 };
 static XYZ_32 m_PickUpPositionUW = { 0, -200, -350 };
-
-static void M_SpawnPickupAid(const ITEM *const item)
-{
-    const OBJECT_ID obj_id =
-        Object_GetCognate(item->object_id, g_ItemToInvObjectMap);
-    const OBJECT *const obj = Object_Get(obj_id);
-    const ANIM_FRAME *const frame = obj->frame_base;
-
-    const GAME_VECTOR pos = {
-        .x = item->pos.x + 20 * (Random_GetDraw() - 0x4000) / 0x4000,
-        .y = item->pos.y - ABS(frame->bounds.max.y - frame->bounds.min.y)
-            - 10 * (1 + (Random_GetDraw() - 0x4000) / 0x4000),
-        .z = item->pos.z + 20 * (Random_GetDraw() - 0x4000) / 0x4000,
-        .room_num = item->room_num,
-    };
-
-    const int16_t effect_num = Effect_Create(pos.room_num);
-    if (effect_num != NO_EFFECT) {
-        EFFECT *const effect = Effect_Get(effect_num);
-        effect->room_num = pos.room_num;
-        effect->pos.x = pos.x;
-        effect->pos.y = pos.y;
-        effect->pos.z = pos.z;
-        effect->counter = 0;
-        effect->object_id = O_PICKUP_AID;
-        effect->frame_num = 0;
-    }
-}
 
 static void M_GetItem(int16_t item_num, ITEM *item, ITEM *lara_item)
 {
@@ -192,81 +152,6 @@ static void M_CollisionControlled(
     item->rot.z = rotz;
 }
 
-static void M_Initialise(int16_t item_num)
-{
-    ITEM *const item = Item_Get(item_num);
-    item->priv = (void *)(intptr_t)(-1);
-    if (item->status != IS_INVISIBLE) {
-        Item_AddActive(item_num);
-    }
-}
-
-static void M_HandleSave(ITEM *const item, const SAVEGAME_STAGE stage)
-{
-    if (stage == SAVEGAME_STAGE_AFTER_LOAD) {
-        if (item->status == IS_DEACTIVATED) {
-            const int16_t item_num = Item_GetIndex(item);
-            Item_RemoveDrawn(item_num);
-        }
-    }
-}
-
-static void M_Activate(ITEM *const item)
-{
-    if (item->status == IS_INVISIBLE) {
-        item->touch_bits = 0;
-        item->status = IS_ACTIVE;
-        const int16_t item_num = Item_GetIndex(item);
-        Item_AddActive(item_num);
-    } else {
-        item->status = IS_INVISIBLE;
-        item->flags |= IF_KILLED;
-    }
-}
-
-static void M_Control(int16_t item_num)
-{
-    ITEM *const item = Item_Get(item_num);
-    if (item->status == IS_INVISIBLE || item->status == IS_DEACTIVATED) {
-        Item_RemoveActive(item_num);
-        return;
-    }
-
-    const ITEM *const lara = Lara_GetItem();
-    if (!g_Config.gameplay.enable_pickup_aids || item->fall_speed != 0
-        || lara == nullptr || !Object_Get(O_PICKUP_AID)->loaded) {
-        return;
-    }
-
-    const int32_t distance = Item_GetDistance(lara, &item->pos);
-    if (distance < AID_DIST_MIN || distance > AID_DIST_MAX) {
-        return;
-    }
-
-    int32_t timer = (int32_t)(intptr_t)item->priv;
-    if (timer <= 0
-        || (timer < AID_WAIT_MIN && Random_GetDraw() < AID_WAIT_BREAK_CHANCE)) {
-        M_SpawnPickupAid(item);
-        timer = AID_WAIT_MAX;
-    } else {
-        timer--;
-    }
-
-    item->priv = (void *)(intptr_t)(int32_t)timer;
-}
-
-static void M_Setup(OBJECT *const obj)
-{
-    obj->draw_func = Object_DrawPickupItem;
-    obj->collision_func = Pickup_Collision;
-    obj->save_flags = true;
-    obj->bounds_func = Pickup_Bounds;
-    obj->initialise_func = M_Initialise;
-    obj->handle_save_func = M_HandleSave;
-    obj->activate_func = M_Activate;
-    obj->control_func = M_Control;
-}
-
 void Pickup_Collision(
     const int16_t item_num, ITEM *const lara_item, COLL_INFO *const coll)
 {
@@ -338,37 +223,3 @@ cleanup:
     item->rot.y = roty;
     item->rot.z = rotz;
 }
-
-REGISTER_OBJECT(O_EXPLOSIVE_ITEM, M_Setup)
-REGISTER_OBJECT(O_FLARES_ITEM, M_Setup)
-REGISTER_OBJECT(O_GRENADE_AMMO_ITEM, M_Setup)
-REGISTER_OBJECT(O_GRENADE_ITEM, M_Setup)
-REGISTER_OBJECT(O_HARPOON_AMMO_ITEM, M_Setup)
-REGISTER_OBJECT(O_HARPOON_ITEM, M_Setup)
-REGISTER_OBJECT(O_KEY_ITEM_1, M_Setup)
-REGISTER_OBJECT(O_KEY_ITEM_2, M_Setup)
-REGISTER_OBJECT(O_KEY_ITEM_3, M_Setup)
-REGISTER_OBJECT(O_KEY_ITEM_4, M_Setup)
-REGISTER_OBJECT(O_LARGE_MEDIPACK_ITEM, M_Setup)
-REGISTER_OBJECT(O_LEADBAR_ITEM, M_Setup)
-REGISTER_OBJECT(O_M16_AMMO_ITEM, M_Setup)
-REGISTER_OBJECT(O_M16_ITEM, M_Setup)
-REGISTER_OBJECT(O_MAGNUM_AMMO_ITEM, M_Setup)
-REGISTER_OBJECT(O_MAGNUM_ITEM, M_Setup)
-REGISTER_OBJECT(O_PICKUP_ITEM_1, M_Setup)
-REGISTER_OBJECT(O_PICKUP_ITEM_2, M_Setup)
-REGISTER_OBJECT(O_PISTOL_AMMO_ITEM, M_Setup)
-REGISTER_OBJECT(O_PISTOL_ITEM, M_Setup)
-REGISTER_OBJECT(O_PUZZLE_ITEM_1, M_Setup)
-REGISTER_OBJECT(O_PUZZLE_ITEM_2, M_Setup)
-REGISTER_OBJECT(O_PUZZLE_ITEM_3, M_Setup)
-REGISTER_OBJECT(O_PUZZLE_ITEM_4, M_Setup)
-REGISTER_OBJECT(O_SCION_ITEM_2, M_Setup)
-REGISTER_OBJECT(O_SECRET_1, M_Setup)
-REGISTER_OBJECT(O_SECRET_2, M_Setup)
-REGISTER_OBJECT(O_SECRET_3, M_Setup)
-REGISTER_OBJECT(O_SHOTGUN_AMMO_ITEM, M_Setup)
-REGISTER_OBJECT(O_SHOTGUN_ITEM, M_Setup)
-REGISTER_OBJECT(O_SMALL_MEDIPACK_ITEM, M_Setup)
-REGISTER_OBJECT(O_UZI_AMMO_ITEM, M_Setup)
-REGISTER_OBJECT(O_UZI_ITEM, M_Setup)
